requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
python-dotenv==1.0.0
gunicorn==20.1.0
pytest==7.3.1
